Recduce code duplication by having a new "shared.py"
